UI refinements: calendar today indicator, day modal lists, session view cleanup, linter fixes
- Weekly calendar: avatar-colored today ring, simplified single-circle design, smoother swipe
- Day detail modal: bullet-point exercise list with indent, set counts, swipeable between days
- Session view: navigates to post-workout summary, compact set display, tap-to-edit weights/reps
- Up Next card: full date label, muscle group pills, exercise preview list
- Post-workout: redesigned summary with volume comparison, effort score, best set
- Various linter and formatting fixes across components

+++ app/history.tsx
@@ -228,7 +228,7 @@ export default function HistoryScreen() {
         {selectedDate && (
           <View style={{ paddingHorizontal: 24, marginTop: 16 }}>
             <Text style={{ fontSize: 13, fontWeight: '600', color: theme.textSecondary, marginBottom: 12 }}>
-              {new Date(selectedDate + 'T12:00:00').toLocaleDateString('en-US', { weekday: 'long', month: 'short', day: 'numeric' })}
+              {new Date(selectedDate + 'T12:00:00').toLocaleDateString('en-US', { weekday: 'long', month: 'long', day: 'numeric' })}
             </Text>
             {dayLogsLoading ? (
               <ActivityIndicator color={theme.chrome} />
